Struct const_matrix::matrix::HMatrix

pub struct HMatrix<T: Numeric<T>, const ROWS: usize, const COLS: usize> { /* fields omitted */ }

A matrix which is allocated on the heap.

Implementations

impl<'a, T: Numeric<T>, const ROWS: usize, const COLS: usize> HMatrix<T, ROWS, COLS>

pub fn new(array: [[T; COLS]; ROWS]) -> Self

Creates a new stack-allocated matrix from the given initial values.

pub fn trans(&self) -> HMatrix<T, COLS, ROWS>

Creates a heap-allocated transpose of this matrix.

pub fn to_stack(&self) -> SMatrix<T, ROWS, COLS>

Creates a stack-allocated copy of this heap-allocated matrix.

impl<T: Numeric<T>, const ROWS_LEFT: usize, const COLS_LEFT: usize> HMatrix<T, ROWS_LEFT, COLS_LEFT>

pub fn mul_heapval_to_stack<const COLS_RIGHT: usize>(
&self,
rhs: HMatrix<T, COLS_LEFT, COLS_RIGHT>
) -> SMatrix<T, ROWS_LEFT, COLS_RIGHT>

Multiply a heap matrix with a heap matrix and allocate the result on the stack.

pub fn mul_heapref_to_stack<const COLS_RIGHT: usize>(
&self,
rhs: &HMatrix<T, COLS_LEFT, COLS_RIGHT>
) -> SMatrix<T, ROWS_LEFT, COLS_RIGHT>

Multiply a heap matrix with a heap matrix and allocate the result on the stack.

pub fn mul_stackval_to_stack<const COLS_RIGHT: usize>(
&self,
rhs: SMatrix<T, COLS_LEFT, COLS_RIGHT>
) -> SMatrix<T, ROWS_LEFT, COLS_RIGHT>

Multiply a heap matrix with a stack matrix and allocate the result on the stack.

pub fn mul_stackref_to_stack<const COLS_RIGHT: usize>(
&self,
rhs: &SMatrix<T, COLS_LEFT, COLS_RIGHT>
) -> SMatrix<T, ROWS_LEFT, COLS_RIGHT>

Multiply a heap matrix with a stack matrix and allocate the result on the stack.