Struct const_matrix::matrix::SMatrix

pub struct SMatrix<T: Numeric<T>, const ROWS: usize, const COLS: usize> { /* fields omitted */ }

A matrix which is allocated on the stack. An SMatrix is itself Numeric, so that its elements can also be other stack-allocated matrices.

Implementations

impl<'a, T: Numeric<T>, const ROWS: usize, const COLS: usize> SMatrix<T, ROWS, COLS>

pub fn new(array: [[T; COLS]; ROWS]) -> Self

Creates a new stack-allocated matrix from the given initial values.

pub fn trans(&self) -> SMatrix<T, COLS, ROWS>

Creates a stack-allocated transpose of this matrix.

pub fn to_heap(&self) -> HMatrix<T, ROWS, COLS>

Creates a heap-allocated copy of this stack-allocated matrix.

impl<T: Numeric<T>, const ROWS_LEFT: usize, const COLS_LEFT: usize> SMatrix<T, ROWS_LEFT, COLS_LEFT>

pub fn mul_val_to_heap<const COLS_RIGHT: usize>(
&self,
rhs: SMatrix<T, COLS_LEFT, COLS_RIGHT>
) -> HMatrix<T, ROWS_LEFT, COLS_RIGHT>

Multiply two stack matrices and allocate the result on the heap.

pub fn mul_ref_to_heap<const COLS_RIGHT: usize>(
&self,
rhs: &SMatrix<T, COLS_LEFT, COLS_RIGHT>
) -> HMatrix<T, ROWS_LEFT, COLS_RIGHT>

Multiply two stack matrices and allocate the result on the heap.

impl<T: Numeric<T>, const ROWS_LEFT: usize, const COLS_LEFT: usize> SMatrix<T, ROWS_LEFT, COLS_LEFT>

pub fn mul_val_to_stack<const COLS_RIGHT: usize>(
&self,
rhs: HMatrix<T, COLS_LEFT, COLS_RIGHT>
) -> SMatrix<T, ROWS_LEFT, COLS_RIGHT>

Multiply a stack matrix with a heap matrix and allocate the result on the stack.

pub fn mul_ref_to_stack<const COLS_RIGHT: usize>(
&self,
rhs: &HMatrix<T, COLS_LEFT, COLS_RIGHT>
) -> SMatrix<T, ROWS_LEFT, COLS_RIGHT>

Multiply a stack matrix with a heap matrix and allocate the result on the stack.